backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
openai
google-generativeai
slowapi
python-multipart
